feat: She Chats
