+++ database/db.go
@@ -1,11 +1,11 @@
-package main
+package database
 
 import (
 	"gorm.io/driver/sqlite"
 	"gorm.io/gorm"
 )
 
-func getDBConnection() (*gorm.DB, error) {
+func GetDBConnection() (*gorm.DB, error) {
 	db, err := gorm.Open(sqlite.Open("transactions.db"), &gorm.Config{})
 	if err != nil {
 		return nil, err
