+++ go.mod
@@ -3,6 +3,7 @@ module github.com/kmesiab/chime-ai
 go 1.23.3
 
 require (
+	github.com/sashabaranov/go-openai v1.36.0
 	gorm.io/driver/sqlite v1.5.7
 	gorm.io/gorm v1.25.12
 )
